Format Python samples and move them into junction-python/
