+++ junction-python/pyproject.toml
@@ -12,6 +12,3 @@ build-backend = "maturin"
 
 [tool.ruff]
 exclude = ["*pb2.py", "*pb2_grpc.py"]
-
-[tool.ruff.lint]
-select = ["E4", "E7", "E9", "F", "B", "I"]
