+++ junction-python/samples/grpc_simple_split/Dockerfile
