+++ junction-python/samples/grpc_simple_split/k8s_jct_grpc_server.yml
