+++ junction-python/samples/grpc_simple_split/xds-bootstrap.json
