+++ junction-python/samples/http_simple_split/Dockerfile
