+++ junction-python/samples/http_simple_split/k8s_gateway.yml
