+++ junction-python/samples/http_simple_split/k8s_jct_http_server.yml
